+++ src/traits.rs
@@ -81,6 +81,13 @@ pub trait Attributes: DataType {
     fn as_node_attributes(&self) -> &ua::NodeAttributes;
 }
 
+/// Custom certificate verification.
+///
+/// This is used to implement custom callbacks in [`ua::CertificateVerification::custom()`].
+pub trait CustomCertificateVerification {
+    fn verify_certificate(&self, certificate: &ua::ByteString) -> ua::StatusCode;
+}
+
 /// Private-key password callback.
 ///
 /// This is used to fetch the password for a given client private key when establishing a connection